Fix: `windDirection` Must Support Values 0[°] to 360[°]

+++ src/BaseWeatherChannel.h
@@ -13,7 +13,7 @@ struct CurrentWheatherData
     uint16_t pressure = 0;
     float windSpeed = 0;
     float windGust = 0;
-    uint8_t windDirection = 0;
+    uint16_t windDirection = 0;
     float rain = 0;
     float snow = 0;
     float uvi = 0;
@@ -28,7 +28,7 @@ struct ForecastHourWheatherData
     uint16_t pressure = 0;
     float windSpeed = 0;
     float windGust = 0;
-    uint8_t windDirection = 0;
+    uint16_t windDirection = 0;
     uint8_t probabilityOfPrecipitation = 0;
     float rain = 0;
     float snow = 0;
@@ -55,7 +55,7 @@ struct ForecastDayWheatherData
     uint16_t pressure = 0;
     float windSpeed = 0;
     float windGust = 0;
-    uint8_t windDirection = 0;
+    uint16_t windDirection = 0;
     uint8_t probabilityOfPrecipitation = 0;
     float rain = 0;
     float snow = 0;
